requirements.txt
selenium>=4.8.3
selenium-wire>=5.1.0
webdriver-manager>=4.0.1
python-decouple==3.8
firebase-admin==6.1.0
pytz==2023.3
blinker<1.8.0
setuptools 
wheel